Implement custom RPC confirmation prompt
For now only qubes.AuthorizeAdminAccess is customized.

QubesOS/qubes-issues#5853
