+++ rpm_spec/qubes-qrexec.spec.in
@@ -165,6 +165,14 @@ rm -f %{name}-%{version}
 %{python3_sitelib}/qrexec/utils.py
 %{python3_sitelib}/qrexec/server.py
 
+%dir %{python3_sitelib}/qrexec/rpcconfirmation
+%dir %{python3_sitelib}/qrexec/rpcconfirmation/__pycache__
+%{python3_sitelib}/qrexec/rpcconfirmation/__pycache__/*
+%{python3_sitelib}/qrexec/rpcconfirmation/__init__.py
+%{python3_sitelib}/qrexec/rpcconfirmation/base.py
+%{python3_sitelib}/qrexec/rpcconfirmation/generic.py
+%{python3_sitelib}/qrexec/rpcconfirmation/admin_authz.py
+
 %dir %{python3_sitelib}/qrexec/policy
 %dir %{python3_sitelib}/qrexec/policy/__pycache__
 %{python3_sitelib}/qrexec/policy/__pycache__/*
@@ -229,6 +237,7 @@ rm -f %{name}-%{version}
 
 %dir %{python3_sitelib}/qrexec/glade
 %{python3_sitelib}/qrexec/glade/RPCConfirmationWindow.glade
+%{python3_sitelib}/qrexec/glade/AdminAuthzRPCConfirmationWindow.glade
 
 %{_datadir}/man/man1/qubes-core-qrexec.1.gz
